Move CredentialsProviderFactory to the public API
